relocating massenergize_errors and route_handler
