+++ src/api/handlers/action.py
@@ -1,6 +1,6 @@
 """Handler file for all routes pertaining to actions"""
 
-from _main_.route_handler import RouteHandler
+from _main_.utils.route_handler import RouteHandler
 from _main_.utils.common import get_request_contents, parse_list, parse_bool, check_length
 from api.services.action import ActionService
 from _main_.utils.massenergize_response import MassenergizeResponse
