+++ src/api/handlers/community.py
@@ -1,6 +1,6 @@
 """Handler file for all routes pertaining to communities"""
 
-from _main_.route_handler import RouteHandler
+from _main_.utils.route_handler import RouteHandler
 from _main_.utils.common import get_request_contents, parse_location, parse_bool, check_length, rename_field
 from api.services.community import CommunityService
 from _main_.utils.massenergize_response import MassenergizeResponse
