+++ src/api/handlers/event.py
@@ -1,6 +1,6 @@
 """Handler file for all routes pertaining to events"""
 
-from _main_.route_handler import RouteHandler
+from _main_.utils.route_handler import RouteHandler
 from _main_.utils.common import get_request_contents, parse_list, parse_bool, check_length, parse_date, parse_int
 from api.services.event import EventService
 from _main_.utils.massenergize_response import MassenergizeResponse
