+++ src/api/handlers/page_settings__contactus.py
@@ -1,6 +1,6 @@
 """Handler file for all routes pertaining to contact_us_page_settings"""
 
-from _main_.route_handler import RouteHandler
+from _main_.utils.route_handler import RouteHandler
 from _main_.utils.common import get_request_contents, rename_field
 from api.services.page_settings__contactus import ContactUsPageSettingsService
 from _main_.utils.massenergize_response import MassenergizeResponse
