+++ src/api/handlers/vendor.py
@@ -1,6 +1,6 @@
 """Handler file for all routes pertaining to vendors"""
 
-from _main_.route_handler import RouteHandler
+from _main_.utils.route_handler import RouteHandler
 from _main_.utils.common import get_request_contents, rename_field, parse_bool, parse_location, parse_list
 from api.services.vendor import VendorService
 from _main_.utils.massenergize_response import MassenergizeResponse
