+++ src/api/services/community.py
@@ -1,4 +1,4 @@
-from _main_.massenergize_errors import MassEnergizeAPIError
+from _main_.utils.massenergize_errors import MassEnergizeAPIError
 from _main_.utils.massenergize_response import MassenergizeResponse
 from api.store.community import CommunityStore
 from _main_.utils.common import serialize, serialize_all
