+++ src/api/store/event.py
@@ -1,5 +1,5 @@
 from database.models import Event, UserProfile, EventAttendee, Media
-from _main_.massenergize_errors import MassEnergizeAPIError, InvalidResourceError, ServerError, CustomMassenergizeError
+from _main_.utils.massenergize_errors import MassEnergizeAPIError, InvalidResourceError, ServerError, CustomMassenergizeError
 from _main_.utils.massenergize_response import MassenergizeResponse
 from django.db.models import Q
