+++ src/api/store/page_settings__actions.py
@@ -1,5 +1,5 @@
 from database.models import ActionsPageSettings, UserProfile
-from _main_.massenergize_errors import MassEnergizeAPIError, InvalidResourceError, ServerError, CustomMassenergizeError
+from _main_.utils.massenergize_errors import MassEnergizeAPIError, InvalidResourceError, ServerError, CustomMassenergizeError
 from _main_.utils.massenergize_response import MassenergizeResponse
 
 class ActionsPageSettingsStore:
